refactor: move component wise codemod logic into separate files
